+++ tango/integrations/torch/train.py
@@ -343,7 +343,7 @@ class TorchTrainStep(Step):
             self.logger.info(
                 f"Loading best weights from {str(config.final_weights_path.resolve())}"
             )
-            state = torch.load(config.final_weights_path, map_location="cpu")
+            state = torch.load(config.final_weights_path, map_location="cpu", weights_only=False)
             # We use `strict=False` because there might be missing keys due to weight tying.
             final_model.load_state_dict(state, strict=False)